requirements.txt
py-cord
aiohttp
PyNaCl
local-ffmpeg
python-dotenv